refactor: blocks routes
